# docs/backend/api/openapi.json
{"openapi":"3.1.0","info":{"title":"VocabVoyage","version":"0.1.0"},"paths":{"/set_mode/":{"post":{"summary":"Set Mode","description":"Sets the quiz mode.\n\nParameters\n----------\nrequest : ModeRequest\n    The request containing the desired mode.\n\nRaises\n------\nHTTPException\n    If the mode is not 'normal' or 'infinite'.\n\nReturns\n-------\nJSONResponse\n    A response indicating the mode has been set.","operationId":"set_mode_set_mode__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/start_quiz/":{"post":{"summary":"Start Quiz","description":"Endpoint to start a new quiz session.\n\nThis endpoint resets the current quiz session and starts a new one. It clears any\nprevious quiz data and prepares the system for a new quiz session. This can be \nuseful when a user wants to restart the quiz from the beginning.\n\nReturns\n-------\n    JSONResponse: A response indicating that the quiz has started.","operationId":"start_quiz_start_quiz__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/end_quiz/":{"post":{"summary":"End Quiz","description":"Endpoint to end the current quiz session and log the results.\n\nThis endpoint marks the end of the current quiz session. It triggers the logging\nof the quiz results, including the number of correct and incorrect answers, and \nany other relevant statistics. This can be useful for tracking user performance \nand providing feedback.\n\nReturns\n-------\n    JSONResponse: A response indicating that the quiz has ended and results have been logged.","operationId":"end_quiz_end_quiz__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/words/next":{"get":{"summary":"Get Next Word","description":"Endpoint to retrieve the next word in the quiz.\n\nReturns\n-------\n    Optional[Word]: The next word in the quiz, or None if there are no more words.","operationId":"get_next_word_words_next_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/Word"},{"type":"null"}],"title":"Response Get Next Word Words Next Get"}}}}}}},"/check/":{"post":{"summary":"Check Answer","description":"Check if the user's answer is correct.\n\nParameters\n----------\nanswer : AnswerRequest\n    The user's answer request containing the word and user input.\n\nReturns\n-------\ndict\n    A dictionary with a key 'is_correct' indicating whether the user's answer is correct.","operationId":"check_answer_check__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Check Answer Check  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/results/":{"get":{"summary":"Get Results","description":"Endpoint to get current quiz statistics.","operationId":"get_results_results__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Results Results  Get"}}}}}}},"/upload_words/":{"post":{"summary":"Upload Words","description":"Endpoint to upload new word files.\n\nThis endpoint allows users to upload new word files in CSV format. It clears any\nexisting word files in the data directory and replaces them with the newly uploaded\nfiles. The words from the new files are then loaded into the system.\n\nParameters\n----------\nfiles : List[UploadFile]\n    A list of uploaded files. Each file must be in CSV format.\n\nRaises\n------\nHTTPException\n    If any of the uploaded files is not a CSV file.\n\nReturns\n-------\ndict\n    A message indicating that the word files have been uploaded and the word list has been updated.","operationId":"upload_words_upload_words__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_words_upload_words__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnswerRequest":{"properties":{"word":{"$ref":"#/components/schemas/Word"},"user_input":{"type":"string","title":"User Input"}},"type":"object","required":["word","user_input"],"title":"AnswerRequest","description":"Request model for submitting an answer.\n\nAttributes\n----------\nword : Word\n    The word object of the questioner.\nuser_input : str\n    The user's input for the word."},"Body_upload_words_upload_words__post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_upload_words_upload_words__post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ModeRequest":{"properties":{"mode":{"type":"string","title":"Mode"}},"type":"object","required":["mode"],"title":"ModeRequest","description":"Request model for setting the quiz mode.\n\nAttributes\n----------\n    mode (str): The quiz mode, either 'normal' or 'infinite'."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Word":{"properties":{"foreign_term":{"type":"string","title":"Foreign Term"},"native_translation":{"type":"string","title":"Native Translation"}},"type":"object","required":["foreign_term","native_translation"],"title":"Word","description":"Represents a word with its foreign term and native translation.\n\nAttributes:\n    foreign_term (str): The word in the foreign language.\n    native_translation (str): The translation of the foreign term in the native language."}}}}